tools/video_ai/requirements.txt
# Video AI Tools Dependencies
# Install with: pip install -r requirements.txt

# Core
pillow>=10.0.0
pytesseract>=0.3.10

# Optional: For training baseline model
# torch>=2.0.0
# torchvision>=0.15.0
# onnx>=1.14.0
# onnxruntime>=1.15.0

# Optional: For FastAPI inference service
# fastapi>=0.100.0
# uvicorn>=0.23.0

# Utilities
tqdm>=4.65.0
